+++ itachallenge-auth/src/main/java/com/itachallenge/auth/App.java
@@ -3,9 +3,11 @@ package com.itachallenge.auth;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
 @EnableDiscoveryClient
+@ComponentScan(basePackages = "com.itachallenge.jwtcore")
 public class App {
     public static void main(String[] args) {
         SpringApplication.run(App.class, args);