+++ itachallenge-auth/src/test/java/com/itachallenge/auth/config/TestAuthConfig.java
@@ -1,0 +1,11 @@
+package com.itachallenge.auth.config;
+
+
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@TestConfiguration
+@ComponentScan(basePackages = "com.itachallenge.auth")
+public class TestAuthConfig {
+}