+++ itachallenge-auth/src/test/java/com/itachallenge/auth/controller/AuthControllerTest.java
@@ -1,25 +1,25 @@
 package com.itachallenge.auth.controller;
 
+import com.itachallenge.auth.config.TestAuthConfig;
 import com.itachallenge.auth.dto.SwitchRoleRequest;
 import com.itachallenge.auth.dto.User;
+import com.itachallenge.auth.service.JwtRoleSwitchService;
 import com.itachallenge.auth.exception.InvalidRoleChangeRequestException;
 import com.itachallenge.auth.service.IAuthService;
-import com.itachallenge.auth.service.IJwtService;
+import com.itachallenge.auth.service.IAuthJwtFacade;
 import com.itachallenge.auth.service.IUserService;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.env.Environment;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.reactive.server.WebTestClient;
-import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Mono;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
@@ -30,8 +30,9 @@ import java.util.Map;
 
 import static org.mockito.Mockito.when;
 
-@ExtendWith(MockitoExtension.class)
 @WebFluxTest(AuthController.class)
+@TestPropertySource(properties = "token.expiration.minutes=60")
+@Import(TestAuthConfig.class)
 @ActiveProfiles("test")
 class AuthControllerTest {
     @Autowired
@@ -44,13 +45,16 @@ class AuthControllerTest {
     private IAuthService authService;
 
     @MockBean
-    private IUserService userService;
+    private JwtRoleSwitchService jwtRoleSwitchService;
 
     @MockBean
-    private IJwtService jwtService;
+    private IUserService userService;
 
     @InjectMocks
     private AuthController authController;
+
+    @MockBean
+    private IAuthJwtFacade authJwtFacade;
 
     @Test
     void authenticateWithGithub_ValidCode_ReturnsJwt() {
@@ -68,7 +72,7 @@ class AuthControllerTest {
         when(authService.exchangeCodeForToken(validCode)).thenReturn(Mono.just(accessToken));
         when(authService.validateTokenWithGithub(accessToken)).thenReturn(Mono.just(validationResult));
         when(userService.fetchUserData(githubUsername)).thenReturn(Mono.just(user));
-        when(jwtService.generateToken(user.getUsername(), user.getRole(), user.getUuid())).thenReturn(jwtToken);
+        when(authJwtFacade.generateToken(user.getUsername(), user.getRole(), user.getUuid())).thenReturn(jwtToken);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/github/authenticate")
@@ -214,7 +218,7 @@ class AuthControllerTest {
     @Test
     void logout_ValidToken_ShouldReturn200() {
         String token = "valid.jwt.token";
-        Mockito.doNothing().when(jwtService).validateToken(token);
+        Mockito.doNothing().when(authJwtFacade).validateToken(token);
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/logout")
                 .header("Authorization", "Bearer " + token)
@@ -228,7 +232,7 @@ class AuthControllerTest {
     void logout_ExpiredToken_ShouldReturn200() {
         String expiredToken = "expired.jwt.token";
         Mockito.doThrow(new ExpiredJwtException(null, null, "Token expired but logout successful"))
-                .when(jwtService).validateToken(expiredToken);
+                .when(authJwtFacade).validateToken(expiredToken);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/logout")
@@ -244,7 +248,7 @@ class AuthControllerTest {
     @Test
     void logout_TokenJustWithinTry_ShouldReturn200() {
         String token = "any.jwt.token";
-        Mockito.doNothing().when(jwtService).validateToken(token);
+        Mockito.doNothing().when(authJwtFacade).validateToken(token);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/logout")
@@ -259,7 +263,7 @@ class AuthControllerTest {
     void logout_InvalidToken_ShouldReturn401() {
         String invalidToken = "invalid.jwt.token";
         Mockito.doThrow(new JwtException("Invalid or tampered token: JWT parsing failed"))
-                .when(jwtService).validateToken(invalidToken);
+                .when(authJwtFacade).validateToken(invalidToken);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/logout")
@@ -285,7 +289,7 @@ class AuthControllerTest {
                 });
     }
 
-   @Test
+    @Test
     void logout_MalformedAuthorizationHeader_ShouldReturn401() {
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/logout")
@@ -316,8 +320,8 @@ class AuthControllerTest {
         String newRole = "ADMIN";
         String newToken = "new.jwt.token";
 
-        when(jwtService.extractBearerToken("Bearer " + originalToken)).thenReturn(originalToken);
-        when(jwtService.switchRole(originalToken, newRole)).thenReturn(newToken);
+        when(authJwtFacade.extractBearerToken("Bearer " + originalToken)).thenReturn(originalToken);
+        when(jwtRoleSwitchService.switchRole(originalToken, newRole)).thenReturn(newToken);
 
         webTestClient.post()
                 .uri("/itachallenge/api/v1/auth/switch-role")
@@ -334,8 +338,8 @@ class AuthControllerTest {
         String expiredToken = "expired.jwt.token";
         String newRole = "ADMIN";
 
-        when(jwtService.extractBearerToken("Bearer " + expiredToken)).thenReturn(expiredToken);
-        when(jwtService.switchRole(expiredToken, newRole))
+        when(authJwtFacade.extractBearerToken("Bearer " + expiredToken)).thenReturn(expiredToken);
+        when(jwtRoleSwitchService.switchRole(expiredToken, newRole))
                 .thenThrow(new ExpiredJwtException(null, null, "Token is expired."));
 
         webTestClient.post()
@@ -353,8 +357,8 @@ class AuthControllerTest {
         String invalidToken = "invalid.jwt.token";
         String newRole = "ADMIN";
 
-        when(jwtService.extractBearerToken("Bearer " + invalidToken)).thenReturn(invalidToken);
-        when(jwtService.switchRole(invalidToken, newRole))
+        when(authJwtFacade.extractBearerToken("Bearer " + invalidToken)).thenReturn(invalidToken);
+        when(jwtRoleSwitchService.switchRole(invalidToken, newRole))
                 .thenThrow(new JwtException("Invalid or tampered token."));
 
         webTestClient.post()
@@ -371,8 +375,8 @@ class AuthControllerTest {
     void switchRole_WithSameRole_Throws400() {
         String token = "valid.jwt.token";
 
-        when(jwtService.extractBearerToken("Bearer " + token)).thenReturn(token);
-        when(jwtService.switchRole(token, "USER"))
+        when(authJwtFacade.extractBearerToken("Bearer " + token)).thenReturn(token);
+        when(jwtRoleSwitchService.switchRole(token, "USER"))
                 .thenThrow(new InvalidRoleChangeRequestException("New role is the same as current role."));
 
         webTestClient.post()
@@ -387,7 +391,7 @@ class AuthControllerTest {
 
     @Test
     void switchRole_MissingAuthorizationHeader_Returns401() {
-        when(jwtService.extractBearerToken(null))
+        when(authJwtFacade.extractBearerToken(null))
                 .thenThrow(new JwtException("Authorization header is missing or malformed"));
 
         webTestClient.post()